backend/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.30
aiosqlite==0.20.0
openpyxl==3.1.2
python-docx==1.1.2
python-pptx==0.6.23
aiofiles==23.2.1
aiosmtplib==3.0.1
pillow>=11.0.0
reportlab>=4.2.0
python-dotenv==1.0.1
